Replace justify-content:center with space-between
Space-between automatically distributes the extra space around logo. Meaning, left and right links are automatically pushed at the start and end of flex container's primary axis.

Padding is also set to 8px in the header so that logo, left, and right-links have the same padding.

Margin and padding of ul is set to 0 so that padding set in header will show consistent look.